Instruction Inputs are now Variables.
To distinguish between a literal and and reference all the inputs and
outputs across the system needed to change to Variables.

This effected across the majority of the system, but has successfully
been tested on MPS code.

Instruction IDs have been changed to long to capture the data from MPS.